Inspector: add `info` function
